- better package structure - rename pages to screens - moving widgets to its own folder - preparing theme file to be used as app theme - update test because of app renaming and refactoring to new file
